+++ examples/magento-graphcms/components/Row/RowServiceOptions/index.tsx
@@ -1,12 +1,13 @@
 import RichText from '@graphcommerce/graphcms-ui/RichText'
 import {
+  IconBlock,
+  IconBlocks,
   iconChat,
   iconEmail,
   iconPhone,
-  IconBlocks,
-  IconBlock,
   SvgImage,
 } from '@graphcommerce/next-ui'
+import { Link } from '@material-ui/core'
 import PageLink from 'next/link'
 import React from 'react'
 import { RowServiceOptionsFragment } from './RowServiceOptions.gql'
@@ -20,6 +21,10 @@ export default function RowServiceOptions(props: RowServiceOptionsProps) {
     <IconBlocks title={title}>
       {serviceOptions.map((serviceOption) => {
         const iconTitle = serviceOption.title.toLowerCase()
+
+        const hrefTitle = serviceOption.url.includes(':')
+          ? serviceOption.url.split(':').pop() ?? ''
+          : serviceOption.url
 
         return (
           <PageLink key={serviceOption.title} href={serviceOption.url} passHref>
@@ -37,6 +42,9 @@ export default function RowServiceOptions(props: RowServiceOptionsProps) {
                 </>
               }
             >
+              <Link href={serviceOption.url} underline='always' color='primary'>
+                {hrefTitle}
+              </Link>
               {serviceOption.description ? <RichText {...serviceOption.description} /> : undefined}
             </IconBlock>
           </PageLink>