+++ packages/framer-scroller/hooks/useScroller.ts
@@ -3,13 +3,12 @@ import { makeStyles } from '@material-ui/core'
 import clsx from 'clsx'
 import {
   HTMLMotionProps,
-  PanInfo,
   motionValue,
-  useDomEvent,
-  PanHandlers,
-  m,
-  useTransform,
   MotionValue,
+  PanHandlers,
+  PanInfo,
+  useDomEvent,
+  useTransform,
 } from 'framer-motion'
 import React, { ReactHTML, useState } from 'react'
 import { ScrollSnapProps } from '../types'
@@ -100,7 +99,10 @@ export function useScroller<TagName extends keyof ReactHTML = 'div'>(
   })
 
   const scrollStart = useConstant(() => ({ x: motionValue(0), y: motionValue(0) }))
-  const onPanStart: PanHandlers['onPanStart'] = () => {
+  const onPanStart: PanHandlers['onPanStart'] = (event) => {
+    // If we're not dealing with the mouse we don't need to do anything
+    if (!isHTMLMousePointerEvent(event)) return
+
     scrollStart.x.set(scroll.x.get())
     scrollStart.y.set(scroll.y.get())
     disableSnap()
@@ -120,6 +122,7 @@ export function useScroller<TagName extends keyof ReactHTML = 'div'>(
   const onPanEnd: PanHandlers['onPanEnd'] = (event, info) => {
     // If we're not dealing with the mouse we don't need to do anything
     if (!isHTMLMousePointerEvent(event)) return
+
     setPanning(false)
     animatePan(info, enableSnap)
   }