+++ packages/magento-cart-payment-method/index.ts
@@ -1,4 +1,5 @@
 export * from './Api/PaymentMethod'
+export * from './hooks'
 
 export * from './PaymentMethodContext/PaymentMethodContext'
 export { default as PaymentMethodContextProvider } from './PaymentMethodContext/PaymentMethodContext'
@@ -15,4 +16,4 @@ export { default as PaymentMethodPlaceOrder } from './PaymentMethodPlaceOrder/Pa
 export { default as PaymentMethodPlaceOrderNoop } from './PaymentMethodPlaceOrderNoop/PaymentMethodPlaceOrderNoop'
 
 export * from './PaymentMethodToggles/PaymentMethodToggles'
-export { default as PaymentMethodToggle } from './PaymentMethodToggles/PaymentMethodToggles'
+export { default as PaymentMethodToggles } from './PaymentMethodToggles/PaymentMethodToggles'