+++ packages/magento-cart/components/CartTotals/CartTotals.tsx
@@ -65,9 +65,10 @@ export default function CartTotals(props: CartTotalsProps) {
   const { containerMargin } = props
   const { shipping_addresses, prices } = data.cart
   const shippingMethod = shipping_addresses?.[0]?.selected_shipping_method
+
   const shippingMethodPrices = shipping_addresses?.[0]?.available_shipping_methods?.find(
     (avail) =>
-      (shippingMethod?.amount.value ?? 0) > 0 &&
+      (shippingMethod?.amount?.value ?? 0) > 0 &&
       avail?.carrier_code === shippingMethod?.carrier_code &&
       avail?.method_code === shippingMethod?.method_code,
   )