+++ packages/magento-cart/hooks/useClearCurrentCartId.ts
@@ -1,14 +1,13 @@
 import { useApolloClient } from '@apollo/client'
 import { CurrentCartIdDocument } from './CurrentCartId.gql'
-import { useCurrentCartId } from './useCurrentCartId'
 
 export function useClearCurrentCartId() {
   const { cache } = useApolloClient()
-  const currentCartId = useCurrentCartId()
-
-  if (!currentCartId) return undefined
 
   return () => {
+    const id = cache.readQuery({ query: CurrentCartIdDocument })?.currentCartId?.id
+    if (!id) return
+
     cache.writeQuery({
       query: CurrentCartIdDocument,
       data: { currentCartId: { __typename: 'CurrentCartId', id: null } },