+++ packages/mollie-magento-payment/methods/mollie_methods_paypal.ts
@@ -1,6 +1,6 @@
 import { PaymentMethodOptionsNoop, PaymentModule } from '@graphcommerce/magento-cart-payment-method'
-import MollieToggle from '../MolliePaymentToggle/MolliePaymentToggle'
-import MolliePlaceOrder from '../MolliePlaceOrder/MolliePlaceOrder'
+import MollieToggle from '../components/MolliePaymentToggle/MolliePaymentToggle'
+import MolliePlaceOrder from '../components/MolliePlaceOrder/MolliePlaceOrder'
 
 export const mollie_methods_paypal: PaymentModule = {
   PaymentToggle: MollieToggle,